+++ test/fake_embulk/output_plugin.rb
@@ -1,0 +1,4 @@
+module Embulk
+  class OutputPlugin
+  end
+end