+++ test/fake_embulk/plugin.rb
@@ -1,0 +1,6 @@
+module Embulk
+  class Plugin
+    def self.register_output(*args)
+    end
+  end
+end