+++ test/fake_embulk/schema.rb
@@ -1,0 +1,15 @@
+module Embulk
+  class Schema < ::Array
+    def initialize(*args)
+      super
+    end
+
+    def names
+      map {|c| c[:name] }
+    end
+
+    def types
+      map {|c| c[:type] }
+    end
+  end
+end